+++ tests/test_log.py
@@ -1,8 +1,11 @@
+# encoding=utf-8
+
+
 import io
 
 from rich.console import Console
 
-from render import render
+from render import replace_link_ids
 
 
 test_data = [1, 2, 3]
@@ -19,11 +22,11 @@ def render_log():
     console.log()
     console.log("Hello from", console, "!")
     console.log(test_data, log_locals=True)
-    return console.file.getvalue()
+    return replace_link_ids(console.file.getvalue())
 
 
 def test_log():
-    expected = "\n\x1b[2;36m[TIME]\x1b[0m\x1b[2;36m \x1b[0mHello from \x1b[1m<\x1b[0m\x1b[1;38;5;13mconsole\x1b[0m\x1b[39m \x1b[0m\x1b[3;33mwidth\x1b[0m\x1b[39m=\x1b[0m\x1b[1;34m80\x1b[0m\x1b[39m ColorSystem.TRUECOLOR\x1b[0m\x1b[1m>\x1b[0m !      \x1b]8;id=1331930600;file://test_log.py\x1b\\\x1b[2mtest_log.py\x1b[0m\x1b]8;;\x1b\\\x1b[2m:20\x1b[0m\n\x1b[2;36m      \x1b[0m\x1b[2;36m \x1b[0m\x1b[1m[\x1b[0m\x1b[1;34m1\x1b[0m, \x1b[1;34m2\x1b[0m, \x1b[1;34m3\x1b[0m\x1b[1m]\x1b[0m                                                  \x1b]8;id=1331930600;file://test_log.py\x1b\\\x1b[2mtest_log.py\x1b[0m\x1b]8;;\x1b\\\x1b[2m:21\x1b[0m\n       \x1b[3m                       Locals                             \x1b[0m               \n       \x1b[34m╭─────────┬────────────────────────────────────────╮\x1b[0m                     \n       \x1b[34m│\x1b[0m\x1b[32m'console'\x1b[0m\x1b[34m│\x1b[0m\x1b[1m<\x1b[0m\x1b[1;38;5;13mconsole\x1b[0m\x1b[39m \x1b[0m\x1b[3;33mwidth\x1b[0m\x1b[39m=\x1b[0m\x1b[1;34m80\x1b[0m\x1b[39m ColorSystem.TRUECOLOR\x1b[0m\x1b[1m>\x1b[0m\x1b[34m│\x1b[0m                     \n       \x1b[34m╰─────────┴────────────────────────────────────────╯\x1b[0m                     \n"
+    expected = "\n\x1b[2;36m[TIME]\x1b[0m\x1b[2;36m \x1b[0mHello from \x1b[1m<\x1b[0m\x1b[1;38;5;13mconsole\x1b[0m\x1b[39m \x1b[0m\x1b[3;33mwidth\x1b[0m\x1b[39m=\x1b[0m\x1b[1;34m80\x1b[0m\x1b[39m ColorSystem.TRUECOLOR\x1b[0m\x1b[1m>\x1b[0m !      \x1b]8;id=0file:///Users/willmcgugan/projects/rich/tests/test_log.py\x1b\\\x1b[2mtest_log.py\x1b[0m\x1b]8;;\x1b\\\x1b[2m:23\x1b[0m\n\x1b[2;36m      \x1b[0m\x1b[2;36m \x1b[0m\x1b[1m[\x1b[0m\x1b[1;34m1\x1b[0m, \x1b[1;34m2\x1b[0m, \x1b[1;34m3\x1b[0m\x1b[1m]\x1b[0m                                                  \x1b]8;id=0file:///Users/willmcgugan/projects/rich/tests/test_log.py\x1b\\\x1b[2mtest_log.py\x1b[0m\x1b]8;;\x1b\\\x1b[2m:24\x1b[0m\n       \x1b[3m                       Locals                             \x1b[0m               \n       \x1b[34m╭─────────┬────────────────────────────────────────╮\x1b[0m                     \n       \x1b[34m│\x1b[0m\x1b[32m'console'\x1b[0m\x1b[34m│\x1b[0m\x1b[1m<\x1b[0m\x1b[1;38;5;13mconsole\x1b[0m\x1b[39m \x1b[0m\x1b[3;33mwidth\x1b[0m\x1b[39m=\x1b[0m\x1b[1;34m80\x1b[0m\x1b[39m ColorSystem.TRUECOLOR\x1b[0m\x1b[1m>\x1b[0m\x1b[34m│\x1b[0m                     \n       \x1b[34m╰─────────┴────────────────────────────────────────╯\x1b[0m                     \n"
     assert render_log() == expected