+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.3] - Unreleased
+
+### Fixed
+
+- Fixed console.log hyperlink not containing full path
+
+### Changed
+
+- Used random number for hyperlink id
+
 ## [2.2.2] - 2020-06-14
 
 ### Changed